+++ assets/scripts/matomo_archive.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
 
 if [ -f /var/www/config/config.ini.php ]; then
-  cd /var/www
-  php console core:archive --url=${SITE_URL}
+  su - nginx -s /bin/sh -c "php /var/www/console core:archive --url=${SITE_URL}"
 fi