+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.3.0-RC2 (2018/01/23)
+
+* Add `bootstrap.php` to move user data in a persistent folder
+* Use `inotifywatch` to check if a plugin installed/removed
+* Preserve plugins across upgrades
+
 ## 3.3.0-RC1 (2018/01/21)
 
 * Initial version